This allows building KaGen as static again, a bug that #74 introduced.
xxhash sets `BUILD_SHARED_LIBS` which forces all libraries to be
build as `SHARED`.

Previously, we fixed this by calling
`set(BUILD_SHARED_LIBS OFF)` before including xxhash, but this cannot
be overwritten by the user. This prevents user from building KaGen as
a shared library, as noticed in #74.

This change no handles that cleanly and gives full control to the user
by add `option(BUILD_SHARED_LIBS ...)` at the top of KaGen's
CMakeList.txt, as suggested in [1].

[1]: https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html